feat: created book genres model, forms and controllers

+++ app/modules/books/controllers.py
@@ -53,6 +53,7 @@ mod_author = Blueprint('authors', __name__, url_prefix='/authors')
 mod_publisher = Blueprint('publishers', __name__, url_prefix='/publishers')
 mod_genre = Blueprint('genres', __name__, url_prefix='/genres')
 mod_book = Blueprint('books', __name__, url_prefix='/books')
+mod_book_genre = Blueprint('book_genres', __name__, url_prefix='/book_genres')
 
 # Set the route and accepted methods
 @mod_author.route('', methods=['GET'])
@@ -588,7 +589,10 @@ def delete_publisher(id):
             # If the item is found
             if item:
                 # Checking if there are relationships defined for the item
-                # TODO: check for associated items
+                if Book.query.filter(Book.publisher_id==id).first() is not None:
+                    return jsonify({"data": [],
+                                    "meta": {"success": False,
+                                            "errors": _("There are other items associated with this item")}}), 400
                 try:
                     session.delete(item)
                     session.commit()
@@ -1154,6 +1158,246 @@ def delete_book(id):
                     if item.photo_url: remove_file(item.photo_url)
                     if item.photo_thumbnail_url: remove_file(item.photo_thumbnail_url)
                     # Removing the item
+                    session.delete(item)
+                    session.commit()
+                    return jsonify({"data": '',
+                                    "meta": {"success": True}}), 204
+                # If an error occurrs
+                except Exception as e:
+                    session.rollback()
+                    return jsonify({"data": [],
+                                    "meta": {"success": False,
+                                            "errors": str(e)}}), 500
+
+            # If no item is found
+            return jsonify({"data": [],
+                                "meta": {"success": False,
+                                        "errors": _("No item found")}}), 400
+
+# Set the route and accepted methods
+@mod_book_genre.route('', methods=['GET'])
+@ensure_authenticated
+def index_book_genre():
+    # For GET method
+    if request.method == 'GET':
+        # Pagination
+        page = request.args.get('page', default=1, type=int)
+        limit = request.args.get('limit', default=25, type=int)
+        # Setting up a maximum number of results per page, even if limit exceeds it
+        max_per_page = 250
+        # Filtering and sorting
+        filter = request.args.get('filter', default='[]', type=str)
+        sort = request.args.get('sort', default='[]', type=str)
+        # Query timezone
+        timezone = request.args.get('timezone', default=os.getenv('TZ', 'UTC'), type=str)
+        try: q_tz = pytz.timezone(timezone)
+        except: q_tz = pytz.timezone(os.getenv('TZ', 'UTC'))
+
+        # Defining the class for the data model, must be updated for different models
+        model = BookGenre
+        selectinloads = eval(''.join(f'selectinload({r}), ' for r in list(model.__mapper__.relationships)))
+
+        # Trying to obtain data from models
+        try:
+            # Retrieving the sorting attributes
+            sort_attrs = get_sort_attrs(model, sort)
+            # Retrieving the join relationship models
+            join_attrs = get_join_attrs(model, filter)
+            # Retrieving the filtering attributes
+            filter_attrs = get_filter_attrs(model, filter, q_tz)
+
+            # Searching itens by filters and sorting
+            if (len(join_attrs) > 0):
+                # If joins are required
+                res = model.query.options(selectinloads).join(*join_attrs).filter(
+                    *filter_attrs).order_by(*sort_attrs).paginate(page, limit, False, max_per_page)
+            else:
+                # If joins are not required
+                res = model.query.options(selectinloads).filter(*filter_attrs).order_by(
+                    *sort_attrs).paginate(page, limit, False, max_per_page)
+            data = [r.as_dict(q_tz)
+                    for r in res.items] if len(res.items) > 0 else []
+
+            # Returning data and meta
+            return jsonify({"data": data,
+                            "meta": {"success": True,
+                                     "count": res.total}})
+        # If something goes wrong
+        except Exception as e:
+            return jsonify({"data": {},
+                            "meta": {"success": False,
+                                     "errors": str(e)}}), 500
+
+# Set the route and accepted methods
+@mod_book_genre.route('', methods=['POST'])
+@ensure_authorized
+def create_book_genre():
+    # For POST method
+    if request.method == 'POST':
+        # Creating the session for database communication
+        with AppSession() as session:
+            # If data form is submitted
+            form = CreateBookGenreForm.from_json(request.json)
+
+            # If something goes wrong when validating provided data
+            if not form.validate():
+                # Returning the data to the request
+                return jsonify({"data": [],
+                                "meta": {"success": False,
+                                        "errors": form.errors}}), 400
+
+            # Checking if field is already in use
+            if session.query(BookGenre).filter_by(book_id=form.book_id.data, genre_id=form.genre_id.data).first():
+                return jsonify({"data": [],
+                                "meta": {"success": False,
+                                        "errors": _("This genre is already in use for this book.")}}), 400
+            try:
+                # Checking if book exists
+                if session.query(Book).get(form.book_id.data) is None:
+                    return jsonify({"data": [],
+                                    "meta": {"success": False,
+                                            "errors": _("No book found")}}), 400
+                # Checking if genre exists
+                if session.query(Genre).get(form.genre_id.data) is None:
+                    return jsonify({"data": [],
+                                    "meta": {"success": False,
+                                            "errors": _("No genre found")}}), 400
+                # Creating new item
+                item = BookGenre(
+                    book_id=form.book_id.data,
+                    genre_id=form.genre_id.data,
+                    )
+                session.add(item)
+                session.flush()
+                session.commit()
+                return jsonify({"data": item.as_dict(),
+                                "meta": {"success": True}})
+            # If an error occurrs
+            except Exception as e:
+                session.rollback()
+                return jsonify({"data": [],
+                                "meta": {"success": False,
+                                        "errors": str(e)}}), 500
+
+# Set the route and accepted methods
+@mod_book_genre.route('/<int:id>', methods=['GET'])
+@ensure_authenticated
+def get_book_genre_by_id(id):
+    # For GET method
+    if request.method == 'GET':
+        # Creating the session for database communication
+        with AppSession() as session:
+            # Getting the model
+            model = BookGenre
+            selectinloads = eval(''.join(f'selectinload({r}), ' for r in list(model.__mapper__.relationships)))
+
+            # Searching item by ID
+            item = session.query(model).options(selectinloads).get(id)
+
+            # If item is found
+            if item:
+                return jsonify({"data": item.as_dict(),
+                                "meta": {"success": True}})
+            
+            # If no item is found
+            return jsonify({"data": [],
+                                "meta": {"success": False,
+                                        "errors": _("No item found")}}), 400
+
+# Set the route and accepted methods
+@mod_book_genre.route('/<int:id>', methods=['PUT'])
+@ensure_authorized
+def update_book_genre(id):
+    # For PUT method
+    if request.method == 'PUT':
+        # Creating the session for database communication
+        with AppSession() as session:
+            # If data form is submitted
+            form = UpdateBookGenreForm.from_json(request.json)
+
+            # If something goes wrong when validating provided data
+            if not form.validate():
+                # Returning the data to the request
+                return jsonify({"data": [],
+                                "meta": {"success": False,
+                                        "errors": form.errors}}), 400
+
+            try:
+                # Updating the item
+                item = session.query(BookGenre).get(id)
+                
+                # If no item is found
+                if item is None:
+                    return jsonify({"data": [],
+                                    "meta": {"success": False,
+                                            "errors": _("No item found")}}), 400
+
+                # Checking if provided data is already in use
+                if form.book_id.data and form.genre_id.data:
+                    if form.book_id.data != item.book_id or form.genre_id.data != item.genre_id:
+                        res = session.query(BookGenre).filter_by(
+                                book_id=form.book_id.data,
+                                genre_id=form.genre_id.data
+                            ).first()
+                        # Checking if field is not already in use
+                        if session.query(BookGenre).filter_by(
+                                book_id=form.book_id.data,
+                                genre_id=form.genre_id.data
+                            ).first():
+                            return jsonify({"data": [],
+                                "meta": {"success": False,
+                                        "errors": _("This genre is already in use for this book.")}}), 400
+                
+                # Checking if book exists
+                if form.book_id.data and session.query(Book).get(form.book_id.data) is None:
+                    return jsonify({"data": [],
+                                    "meta": {"success": False,
+                                            "errors": _("No book found")}}), 400
+                # Checking if genre exists
+                if form.genre_id.data and session.query(Genre).get(form.genre_id.data) is None:
+                    return jsonify({"data": [],
+                                    "meta": {"success": False,
+                                            "errors": _("No genre found")}}), 400
+                
+                item.book_id = form.book_id.data or item.book_id
+                item.genre_id = form.genre_id.data or item.genre_id
+                
+                try:
+                    session.commit()
+                    return jsonify({"data": item.as_dict(),
+                                    "meta": {"success": True}})
+
+                # If something goes wrong while committing
+                except Exception as e:
+                    session.rollback()
+                    # Returning the data to the request
+                    return jsonify({"data": [],
+                                    "meta": {"success": False,
+                                            "errors": str(e)}}), 500
+
+            # If an error occurrs
+            except Exception as e:
+                session.rollback()
+                return jsonify({"data": [],
+                                "meta": {"success": False,
+                                        "errors": str(e)}}), 500
+
+# Set the route and accepted methods
+@mod_book_genre.route('/<int:id>', methods=['DELETE'])
+@ensure_authorized
+def delete_book_genre(id):
+    # For DELETE method
+    if request.method == 'DELETE':
+        # Creating the session for database communication
+        with AppSession() as session:
+            # Searching item by ID
+            item = session.query(BookGenre).filter_by(id=id).first()
+
+            # If the item is found
+            if item:
+                # Checking if there are relationships defined for the item
+                # TODO: check for associated items
+                try:
                     session.delete(item)
                     session.commit()
                     return jsonify({"data": '',
